refactor: preview step of SessionLauncher into a separate component (#2980)
Resolves #2981

This PR refactors the final step of the SessionLauncher component into a separate component. This change enhances code modularity and maintainability, enabling easier updates and independent testing of the final step.

### Changes
- Extracted the code related to the final step from SessionLauncher.
- Created a new component for the final step.
- Integrated the new component with the existing SessionLauncher.
