+++ react/src/components/BAICard.tsx
